+++ src/Vortice.DXGI/WinUI/ISwapChainPanelNative.cs
@@ -1,0 +1,41 @@
+// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using SharpGen.Runtime;
+
+namespace Vortice.DXGI.WinUI
+{
+    [Guid("63aad0b8-7c24-40ff-85a8-640d944cc325")]
+    public unsafe partial struct ISwapChainPanelNative
+    {
+        public void** lpVtbl;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public int QueryInterface(Guid* riid, void** ppvObject)
+        {
+            return ((delegate* stdcall<ISwapChainPanelNative*, Guid*, void**, int>)(lpVtbl[0]))((ISwapChainPanelNative*)Unsafe.AsPointer(ref this), riid, ppvObject);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public uint AddRef()
+        {
+            return ((delegate* stdcall<ISwapChainPanelNative*, uint>)(lpVtbl[1]))((ISwapChainPanelNative*)Unsafe.AsPointer(ref this));
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public uint Release()
+        {
+            return ((delegate* stdcall<ISwapChainPanelNative*, uint>)(lpVtbl[2]))((ISwapChainPanelNative*)Unsafe.AsPointer(ref this));
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public Result SetSwapChain(IDXGISwapChain swapChain)
+        {
+            void* swapChainPtr = swapChain.NativePointer.ToPointer();
+            return ((delegate* stdcall<ISwapChainPanelNative*, void*, int>)(lpVtbl[3]))((ISwapChainPanelNative*)Unsafe.AsPointer(ref this), swapChainPtr);
+        }
+    }
+}